move build_amount to test_amount
